GUI: Refactoring progress/finished handling.
Progress signal had also item count with it and then the handler determined
that check is ready when max count of progress was done. Also progressbar
was practically reset in every progress signal. This was simply fragile
code.

After this patch progress signal has only the current progress count. Total
count of items is given when initializing the checking. And there is
separate function for handling check finishing.

This also fixes the bug that progressbar was not hidden after checking
or when interrupting the checking.