Correction for flags setting in CMake scripts (bug #2524)
Additional compilation flags were passed to the command "set" as a second parameter.
This had the effect that the passed parameters were converted to a CMake list.
The variable "CMAKE_CXX_FLAGS" got an unwanted semicolon if it was not empty before.

These flags are handled by a single string now.

Signed-off-by: Markus Elfring <elfring@users.sourceforge.net>