GUI: Don't show error when no application added.
I earlier added a check for missing (editor) application command
line parameters. If parameters are missing at GUI startup an
message is shown to user to check the settings. Unfortunately
my check was incomplete and it also triggered when there was no
applications defined. Which happens when user starts GUI for a
first time.

Fix the check so that user is only informed in case the parameters
are missing and name + path are set.