GUI: Remeber last path where project file was opened from.
It is handy to remember the last location of the opened project file.
Currently the Open Project -dialog was always opened to location of
the executable file. Which is never the correct place. But last
opened project file location might at least be near the location user
wants to open next.

Ticket: #3493 (GUI: remember last path in Open Project File)