generate version information by git tags.
This removes the step to manually adapt the version information in
lib/cppcheck.cpp but the Makefile calls `git describe --tags` which yields
the version currently being used.

The format of `git describe` is either
the tag name if a tag is checked out or a string like 1.57-9-g93f1fed
which has the latest reachable tag at first (1.57 in this case). The
information is the number of commits since that tag (we have 9 commits
since that commit) and at last the commit id (behind the 'g': 93f1fed
is the commit before this commit as of now).