Fix for TST-418.
cppcheck sees the function signitures in the connect macro and
as it doesn't have visibility of the Qt macros assumes that
the connect() is calling those functions rather than passing
a reference to the functions.

This fix checks to if calls in the constructor are prefixed
with SLOT or SIGNAL and if so they are ignored.