Fix whitespace caused by MSVC.
Add code to handle QML code blocks (onClicked) & Q_PROPERTYs in C++.

Replace the first block in unused function with non-qt specific code.

Whitespace. Again.

Remove the qt option from settings.cpp

Remove the QT part of the documentation.

Fix whitespace from MSVC.

Fix whitespace from MSVC.

Fix the whitespaces from MSVC.

Fix building under mingw & whitespace.

Add a hook to the ignore function in the Pure Virtual check.

cppcheck, without the QT headers (which are huge) will report that for the following:

class A {
signal:
     virtual x() = 0;
};

class B {
public:
    B(A * a) {
        connect(a, SIGNAL(x()), SLOT(y()); // ERROR, call to pure virtual A.x() in constructor.
    }
slot:
     y();
};

This is fine, we're not calling it, we're just connecting it. And we don't have the Qt headers are parsing them all would take too long.

Instead we know that they're fine and mark them as ignored.

Add the ability to avoid reporting errors for library files.

  <files>
    <file ext=".qml" reporterrors="false"/> <!-- not reported -->
    <file ext=".xxp" reporterrors="true"/>  <!-- reported -->
    <file ext=".other"/>                    <!-- reported -->
  </files>

Add support for library files (i.e. Qt QML)

In the configuration XML we can add a list of files that should be supported and scanned.

In this case, QML files are JavaScript with a C-style syntax and we can ignore the errors on them but it'll help us find function usage.

<!-- definition of the extra files that we want to process -->
<files>
  <file ext=".qml"/>
</files>

Recompiled without Qt settings. Still need to move the updated cpphcheck code into the configuration.