Add a hook to the ignore function in the Pure Virtual check.
cppcheck, without the QT headers (which are huge) will report that for the following:

class A {
signal:
     virtual x() = 0;
};

class B {
public:
    B(A * a) {
        connect(a, SIGNAL(x()), SLOT(y()); // ERROR, call to pure virtual A.x() in constructor.
    }
slot:
     y();
};

This is fine, we're not calling it, we're just connecting it. And we don't have the Qt headers are parsing them all would take too long.

Instead we know that they're fine and mark them as ignored.