Add a .mailmap file.
The .mailmap feature is used to coalesce together commits
by the same person in the shortlog, where their name and/or
email address was spelled differently.
  -- "man git shortlog"

As this corrects the author information at the root (this repository),
all services parsing these information will improve by this contribution.
As discussed on IRC, this also removes the need for different aliases
on ohloh.net (currently https://www.ohloh.net/p/cppcheck/aliases)

        # Finding out duplicates by comparing email addresses:
        git shortlog -sne |awk '{ print $NF }' |sort |uniq -d

        # Finding out duplicates by comparing names:
        git shortlog -sne |awk '{ NF--; $1=""; print }' |sort |uniq -d

Signed-off-by: Stefan Beller <stefanbeller@googlemail.com>