gui: Add compile time option for defining the language directory
This commit adds the compile time option "LANGDIR", which is the path
where the language files can be found. If defined, cppcheck-gui will
look for the language files there, instead of the lang/ subdirectory
of the current working directory.

This is useful in any case where cppcheck-gui is not run from a place where
the lang/ directory is present, but it's mainly meant to make the
language files easily accessible when cppcheck-gui is installed system wide.