Add a dynamic rule enable/disable system
First, each rule now has a "state" attribute:

<rule>
	<state>enabled</state> => rule is enabled
	...
</rule>

<rule>
	<state>disabled</state> => rule is disabled.
	...
</rule>

Second, each rule has a list of IDs of rules to
disable and a list of rules to enable if fired:

<rule>
	...
	<disable>thatrule</disable>
	<enable>thisrule</enable>
</rule>

<rule>
	<state>disabled</state>
	<message>
		<id>thisrule</id>
	...

<rule>
	<message>
		<id>thatrule</id>
	...
</rule>

Initially, the first and third rule are enabled. When
the first rule gets fired, then the third rule becomes
disabled and the second rule becomes enabled.

Note that the behaviour of this system is undefined when
the rule to be enabled or disabled should fire on the same
source fragment which fired the controlling rule.