gui: Check CFGDIR if it's defined.
For quite a while, cppcheck-gui hasn't shown any of the standard
library packages on my Gentoo system. It turns out that cppcheck-gui
doesn't use CFGDIR, but it does use a DATADIR variable stored in
QSettings. Problem is, DATADIR isn't set unless you manually specify
--data-dir, which isn't very intuitive.

This commit adds CFGDIR to the default list of cfg paths if the
CFGDIR macro is defined during the build.