std library: Resolve two TODOs
First TODO in std.c is obsolete since the "res" variable is used later
and there is therefore no warning issued.
Second TODO in std.c: A warning for uninit variables is issued by
cppcheck, so this check can be enabled. But to let "make checkcfg"
succeed also the noreturn in std.cfg had to be specified for assert.
Since assert normally does return (depending on the implementation it
always returns and only prints something) it only makes sense to
configure it as noreturn=false.