cppcheck-cfg.rng: Allow return value by reference
I need a solution to configure functions that return a value by
reference.
This PR might not be the best solution, since it also allows reference
to reference to...
I'm not sure if this is valid C++, but generally a reference to a
reference is possible.
I'm also no regex expert, so any help here is greatly apreciated.