Platforms: Add arm32 and arm64 platform files
The default is "unsigned" so i have choosen it for the platformfiles.
This can be changed via a compiler option, but i am not sure if it makes
sense to add all the platform files with "signed" also.
Since the size of wchar_t depends on the compiler i added both possible
variants respectively.
Sizes of data types can be checked here for example:
http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.den0024a/ch08s02.html