GUI: Fix wrong receivers in connect calls
Since the slots (functions) are not existent in the class MainWindow and
Qt complains that it can not connect the slots i guess it has been
forgotten to change the receiver.
The ResultsView class (from which mUI.mResults is an instance) contains
these slots, so i guess they should be connected here.