Added message field to Suppression & ErrorMessage
This is to add support for suppressing a warning using the text of its warning message, plus wildcards.  This is useful for suppressing warnings like knownConditionTrueFalse (which has no symbol) when they are emitted for code within a macro.  There is no other way of doing a specific suppression for this.