MISRA: Restore multi-file checking
Versions prior to 1.84 had the ability to examine multiple files on the
command line.  This feature was removed in the switch to use argparse
for parsing command line arguments.

Restore the ability to check multiple .dump files provided on the
command line.

Make parseDumps() return the number of violations instead of exiting
and move the exit code reporting up to the top level of the script.

When VERIFY is active do not use the length of the VIOLATIONS list
because the meaning of a violation changes.  VERIFY will always generate
violations but they get classified into expected and not expected.  If
either of expected/not expected checks fail then increment the number of
violations.