MISRA: Restore original summary behavior
Version 1.84 introduced a regression in the behavior of the rule summary
output due to changes in the way multiple input files were handled.

The intended behavior of the summary was to output the total number of
violations after all files have been processed.

Commit aa831ce9721c35a80a2b9aa173e169d2b88905be restored the input file
handling behavior but left summary behavior such that a summary output
was produced for each file that caused a violation instead of the total
number of violations after all files were processed.

Move the -verify logic up into the main loop so that the exit calls are
in the top level and restore the original behavior of the summary
output.