MISRA: Support per file rule suppressions
Parse the suppressions list from cppcheck and extract MISRA rule strings from
the suppressions class by matching for errorId strings that begin with
'MISRA' or 'misra'.  Extract the MISRA rule from those strings by
looking for a '_' or a '.' to separate rule numbers.

Store the rule number, filename, line number, and symbol name from the
suppression entry into a structure that allows for dictionary lookups
by the rule number and then the filename.  All the line number
and symbol entries for that filename are are stored in list of tuples of
(line number, symbol name).  A rule entry that has a value of None for
the filename is treated as a global suppression for all files.  A
filename entry that has None for the rule items list is treated as a
suppression for the entire file. If the rule item list exist then it is
searched for matching line numbers.

Although symbol names are parsed and added int the list of rule items
they are not used for rule matching.  Symbol names can include regular
expressions.  Adding support for symbol names and regular expressions is
left as a future feature.

The existing global suppression method provided by the --suppress-rules
option is supported.  Those rules are added into the suppressions
structure as if they were provided by the suppressions list as global
suppressions. ie A rule with a None for the filename value.