Fix template simplifier and symbol database namespace support.
Forward declarations of classes and structs are now saved in the symbol
database as an empty scope(no {}). This allows a forward declaration in
a nested namespace and an implementation in a higher namespace to be
tied together so the implementation ends up where it was forward
declared rather than where it was implemented.  The template simplifier
no longer embeds the namespace in the template name. The namespace will
now be added as seperate tokens so the symbol database can make sense of
the namespace hiearchy.