Fix regression from unexpected empty scope.
When creating scopes we add an empty placeholder scope for forward
declarations.  The empty placeholder scope will be filled in if an
implementatuion is found.

After all scopes have been created the remaining placeholder scopes will
be removed. With the placeholder scopes removed, no checks will need to
be made aware of the placeholder scopes.

This commit removes the remaining placeholder scopes after scope
creation and reverts the changes to the checks that made them aware of
the placeholder scopes.