donate-cpu.py: Swap "head results" and "diff" output.
As mentioned in https://trac.cppcheck.net/ticket/8881 the "diff" output could be truncated or dropped if the "head results" output is very long. This is already a problem for several packages (for example for http://cppcheck.osuosl.org:8000/gdal).
Since i guess that the "diff" output is more important and normally is way smaller than the "head results" this commit just swaps them so the "diff" output comes first. Thus it is very unlikely that the "diff" output is truncated or dropped.