donate-cpu-server.py: Enhance time output to show long running checks.
Currently there are some very long running checks (several hours) for packages that are not that huge (smaller than 2 MB). So i thought it is worth to list those checks that run for a long time to maybe find performance issues when looking at them.
The threshhold is currently set to 30 * 60 seconds (half an hour).