donate-cpu.py: Add option for limiting bandwidth
With this new option it is possible to limit the download rate / bandwidth that wget uses to download packages.
The value is verified by running wget with it and check if there is a parse error (exit code 2). If the value is not valid an error message is shown and no bandwidth limit is shown.