Donate CPU: Add scripts under /test for testing
Indirectly the scripts (at first the server test script) also document what needs to be done to setup a (local) server.
The productive client script can be used by everyone who wants to support daca@home, not only for testing. But i still think it is good to have it under /test to not loose to much clarity / lucidity in the /tools directory.