donate-cpu-server.py: Verify compared Cppcheck version
Ignore normal results (not fast or info) where the diff was made agains the wrong OLD_VERSION. This avoids unwanted results when some client still analyzes an old package but the OLD_VERSION in the server script is changed for example.
Results missing the Cppcheck version info completely are also ignored.
Tested locally with correct and wrong version numbers.