Load std.cfg before all other libraries
CLI: Save the libraries that should be loaded to a list and load them
after the std.cfg has been loaded.
GUI: Load std.cfg (and windows.cfg / posix.cfg when applicable) before
setting other options and loading the other libraries.
In the project-file-dialog the std.cfg is searched first. If some
other library fails to load is is retried with first loading std.cfg.